# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Training Documents API",
    "description": "An API to list & manage alarm panel training/reference documents built for CSE341",
    "version": "1.0.0"
  },
  "host": "training-docs.onrender.com",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/paneldocs/": {
      "get": {
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "manufacturer": {
                  "example": "any"
                },
                "model": {
                  "example": "any"
                },
                "docType": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "URL": {
                  "example": "any"
                },
                "dateAdded": {
                  "example": "any"
                },
                "user": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/users/": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "any"
                },
                "lastName": {
                  "example": "any"
                },
                "userName": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "authLevel": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}